don't use breaking cargo.toml file for some local projects I am working on

+++ src/curves/curve2d_renderer.rs
@@ -273,7 +273,7 @@ impl Curve2DRenderer {
                 normalize: false,
                 kind: AttributeType::Interleaved {
                     stride,
-                    offset: (Curve2DVertexData::ELEM_SIZE * 2) as u8,
+                    offset: Curve2DVertexData::ELEM_SIZE * 2,
                 },
             },
         ];
